requirements.txt
openai==1.21.2
aiogram==3.24.0
python-dotenv==1.0.1
aiosqlite==0.20.0
aiohttp==3.9.5
requests==2.31.0
